refactor: code review kamar-api

+++ Bilreg.Infrastructure/BillContext/RoomChargeSub/KamarAgg/KamarDal.cs
@@ -45,10 +45,10 @@ public class KamarDal : IKamarDal
         dp.AddParam("@fs_ket1",model.Ket1 ,SqlDbType.VarChar);
         dp.AddParam("@fs_ket2",model.Ket2 ,SqlDbType.VarChar);
         dp.AddParam("@fs_ket3",model.Ket3 ,SqlDbType.VarChar);
-        dp.AddParam("@fn_jumlah",model.JumlahKamar ,SqlDbType.Decimal);
-        dp.AddParam("@fn_pakai",model.JumlahKamarPakai ,SqlDbType.Decimal);
-        dp.AddParam("@fn_kotor",model.JumlahKamarKotor ,SqlDbType.Decimal);
-        dp.AddParam("@fn_rusak",model.JumlahKamarRusak ,SqlDbType.Decimal);
+        dp.AddParam("@fn_jumlah",model.JumlahKamar ,SqlDbType.Int);
+        dp.AddParam("@fn_pakai",model.JumlahKamarPakai ,SqlDbType.Int);
+        dp.AddParam("@fn_kotor",model.JumlahKamarKotor ,SqlDbType.Int);
+        dp.AddParam("@fn_rusak",model.JumlahKamarRusak ,SqlDbType.Int);
         dp.AddParam("@fs_kd_bangsal",model.BangsalId ,SqlDbType.VarChar);
         dp.AddParam("@fs_kd_kelas", model.KelasId, SqlDbType.VarChar);
         
@@ -81,10 +81,10 @@ public class KamarDal : IKamarDal
         dp.AddParam("@fs_ket1",model.Ket1 ,SqlDbType.VarChar);
         dp.AddParam("@fs_ket2",model.Ket2 ,SqlDbType.VarChar);
         dp.AddParam("@fs_ket3",model.Ket3 ,SqlDbType.VarChar);
-        dp.AddParam("@fn_jumlah",model.JumlahKamar ,SqlDbType.Decimal);
-        dp.AddParam("@fn_pakai",model.JumlahKamarPakai ,SqlDbType.Decimal);
-        dp.AddParam("@fn_kotor",model.JumlahKamarKotor ,SqlDbType.Decimal);
-        dp.AddParam("@fn_rusak",model.JumlahKamarRusak ,SqlDbType.Decimal);
+        dp.AddParam("@fn_jumlah",model.JumlahKamar ,SqlDbType.Int);
+        dp.AddParam("@fn_pakai",model.JumlahKamarPakai ,SqlDbType.Int);
+        dp.AddParam("@fn_kotor",model.JumlahKamarKotor ,SqlDbType.Int);
+        dp.AddParam("@fn_rusak",model.JumlahKamarRusak ,SqlDbType.Int);
         dp.AddParam("@fs_kd_bangsal",model.BangsalId ,SqlDbType.VarChar);
         dp.AddParam("@fs_kd_kelas",model.KelasId ,SqlDbType.VarChar);
         
@@ -143,40 +143,33 @@ public class KamarDal : IKamarDal
     private static string SelectFromClause() =>
         @"
         SELECT 
-        aa.fs_kd_kamar,
-        aa.fs_nm_kamar,
-        aa.fs_ket1,
-        aa.fs_ket2,
-        aa.fs_ket3,
-        aa.fn_jumlah,
-        aa.fn_pakai,
-        aa.fn_kotor,
-        aa.fn_rusak,
-        aa.fs_kd_bangsal,
-        aa.fs_kd_kelas,
-        ISNULL(bb.fs_nm_bangsal,'') AS fs_nm_bangsal,
-        ISNULL(cc.fs_nm_kelas,'') AS fs_nm_kelas
+            aa.fs_kd_kamar, aa.fs_nm_kamar,
+            aa.fs_ket1, aa.fs_ket2, aa.fs_ket3,
+            aa.fn_jumlah, aa.fn_pakai, aa.fn_kotor, aa.fn_rusak, 
+            aa.fs_kd_bangsal, aa.fs_kd_kelas,
+            ISNULL(bb.fs_nm_bangsal,'') AS fs_nm_bangsal,
+            ISNULL(cc.fs_nm_kelas,'') AS fs_nm_kelas
 
         FROM ta_kamar aa
-        LEFT JOIN ta_bangsal bb ON aa.fs_kd_bangsal = bb.fs_kd_bangsal
-        LEFT JOIN ta_kelas cc ON aa.fs_kd_kelas = cc.fs_kd_kelas
+            LEFT JOIN ta_bangsal bb ON aa.fs_kd_bangsal = bb.fs_kd_bangsal
+            LEFT JOIN ta_kelas cc ON aa.fs_kd_kelas = cc.fs_kd_kelas
     "; 
 
 }
 
 public class KamarDto() : KamarModel(string.Empty,string.Empty)
 {
-    public string fs_kd_kamar{get => KamarId ;set => KamarId =value;}
-    public string fs_nm_kamar{get => KamarName;set => KamarName =value;}
-    public string fs_ket1{get => Ket1;set => Ket1=value;}
-    public string fs_ket2{get => Ket2;set => Ket2 =value;}
-    public string fs_ket3{get => Ket3;set => Ket3=value;}
-    public decimal fn_jumlah{get => JumlahKamar;set => JumlahKamar =value;}
-    public decimal fn_pakai{get => JumlahKamarPakai;set =>JumlahKamarPakai =value;}
-    public decimal fn_kotor{get => JumlahKamarKotor;set => JumlahKamarKotor=value;}
-    public decimal fn_rusak{get => JumlahKamarRusak;set => JumlahKamarRusak=value;}
-    public string fs_kd_bangsal{get => BangsalId;set => BangsalId=value;}
-    public string fs_nm_bangsal{get => BangsalName;set => BangsalName=value;}
+    public string fs_kd_kamar{get => KamarId ;set => KamarId = value;}
+    public string fs_nm_kamar{get => KamarName;set => KamarName = value;}
+    public string fs_ket1{get => Ket1;set => Ket1 = value;}
+    public string fs_ket2{get => Ket2;set => Ket2 = value;}
+    public string fs_ket3{get => Ket3;set => Ket3 = value;}
+    public int fn_jumlah{get => JumlahKamar;set => JumlahKamar = value;}
+    public int fn_pakai{get => JumlahKamarPakai;set =>JumlahKamarPakai = value;}
+    public int fn_kotor{get => JumlahKamarKotor;set => JumlahKamarKotor = value;}
+    public int fn_rusak{get => JumlahKamarRusak;set => JumlahKamarRusak = value;}
+    public string fs_kd_bangsal{get => BangsalId;set => BangsalId = value;}
+    public string fs_nm_bangsal{get => BangsalName;set => BangsalName = value;}
     public string fs_kd_kelas{get => KelasId;set => KelasId=value;}
     public string fs_nm_kelas{get => KelasName;set => KelasName=value;}
 }
